requirements.txt
alpha-vantage==2.3.1
Flask-APScheduler==1.13.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask==2.3.3
mysql-connector-python==8.1.0
mysql-connector==2.2.9
mysqlclient==2.2.0
numpy==1.25.2
pandas==2.0.3
plotly==5.16.1
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy-Utils==0.41.1
tiingo==0.14.0
WTForms==3.0.1
